CFY-6091 Use postgres as a backend to get list of events
In this implementation the code returns the same results as would have
returned with elasticsearch. This is because it's difficult to update
the events/logs structure without breaking some other part of the code.
Once this implementation is stable and elasticsearch has been removed,
changes to the data structure can be attempted.