CFY-6091 Fix select query
It turns out that `.subquery()` creates a subquery in the FROM clause
which is not what is needed in this case (it returns more matches than
it should because it basically joins events with themselves). To create
a subquery in the SELECT clause `.as_scalar()` needs to be used instead.