CFY-6091. Use postgres to get events (#656)
* CFY-6091 Use postgres as a backend to get list of events

In this implementation the code returns the same results as would have
returned with elasticsearch. This is because it's difficult to update
the events/logs structure without breaking some other part of the code.
Once this implementation is stable and elasticsearch has been removed,
changes to the data structure can be attempted.

* CFY-6091 Update test case

The test case now passes, but still needs more updates. Before this
commit, the mock used to get results is the elasticsearch one while now
results should come from the in-memory sqlite database that is created
for the test case.

* CFY-6091 Add query to count total number of records for an execution

The total number of matches was returned by elasticsearch and also seems
to be used by the client to know when to send multiple queries to the
API when all the results don't fit in a single chunk because of the
pagination requested.

* CFY-6091 Move serialize result to a separate method

There is no reason to have it as a closure since it's not using any of
the variables from the closure

* CFY-6091 Add implementation for v1 API

The implementation for events API v1 has been made by querying the SQL
database and returning data that resembles the one returned from
elasticsearch. This way, the client doesn't need to be changed.

* CFY-6091 Use static methods when the self argument is not used

Helper methods that don't really make use of any instance variable
should be made static

* CFY-6091 Reorder imports

This is to make pylint happy regarding writing imports for the same
library together

* CFY-6091 Add documentation strings

* CFY-6091 Remove engine variable

Access it directly from the `db` global

* CFY-6091 Move query events to a separate method

This is because post queries seem to work the same way as get ones for
this endpoint and implementation is reuse. Note that this is only in v1,
in v2 the post method is just not allowed.

* CFY-6091 Remove redundant code

Most of the query code is now present in the Events v1 resource class
from which the Events v2 resource one inherits, so there's no need to
keep to copies of the same methods.

* CFY-6091 Remove outdated test cases

These test cases verify a method that was present in the old
implementation based on an ElasticSearch backend and no longer apply.

In subsequent commits, similar test cases that exercise the SQL backend
should be added.

* CFY-6091 Remove unused mock method

This was used when the backend used for events was elasticsearch instead
of a SQL database

* CFY-6091 Add missing underscore to sort parameter

The underscore prefix is needed for the sortable rest decorator to work
as expected.

* CFY-6091 Remove unused code

That is code that was used when the backend for events was elasticsearch
instead of a SQL database.

* CFY-6091 Add build select query test cases

The test cases basically verify that the input is validated as expected

* CFY-6091 Add build count query test cases

These test cases verify that the filter parameter is validated and the
query is generated as expected

* CFY-6091 Add test cases for the mapping from sql to elasticsearch

This is needed to make sure that the client is still able to decode the
information even if it's not longer comming from elasticsearch.

* CFY-6091 Update select query to use sqlalchemy instead of raw SQL

* CFY-6091 Update count query to use sqlalchemy instead of raw SQL

* CFY-6091 Update delete query to use sqlalchemy instead of raw SQL

* CFY-6091 Use BadParametersError instead of BadRequest exception

When the parameters don't have the expected value, the
`BadParametersError` is the exception that should be used given that is
handled by the `exceptions_handled` decorator to provide a consistent
response

* CFY-6091 Update select queries to get deployment id from deployments table

With these new queries, there's no need to store a reference to a
deployment, since the execution already has it.

* CFY-6091 Update delete events query to not require deployment_fk

Now events that have a reference to an execution which in turn
references a given deployment are deleted. This is not as
straightforward, but doesn't require to have a reference to the
deployment in each event.

* CFY-6091 Remove unused deployment_fk columns

After the update in the queries, there's no longer need to have a
reference to the deployment in the event tables, since it can be accessed
through the execution which already has a reference to the deployment.

* CFY-6091 Fix delete events test case

The delete events endpoint returns a list in which the first element is
the number of elements deleted instead of an empty list. This is
consistent which what the client implementation expects.

* CFY-6091 Do not synchronize session when running bulk delete

This is to avoid the extra `select` query that sqlalchemy needs to run
to know what objects to update in the database. The synchronization is
not needed because the session is expired after the `.commit` call
anyway.