CFY 6332. Fix events integration tests cases (#700)
* Do not require filtering by execution_id

Despite the client always uses an execution_id, the integration tests don't use any execution_id.

* Do not require ordering ascending by timestamp

No ordering in the timestamp field can be set and implementation works
better with the integration test cases.

* CFY-6332 Enable filtering by deployment_id

* CFY-6332 Include @timestamp field in response

The event data coming from Elasticsearch had both `timestamp` and
`@timestamp`. In this commit the `@timestamp` field is added because
some tests still depend on it. In the future, code should be updated to
use only `timestamp`.

* CFY-6332 Fix save query once sorting order is applied

* CFY-6332 Add support for _include request parameter

For now it's implemented in the python side, but ideally the select
query should be updated to let the database do the projection itself.

* CFY-6332 Apply projection only when needed

Take into account that `_include` might be `None` and then no projection
needs to be applied.

* CFY-6332 Update count query to take deployments into account

Given that the select query returns only events that have a deployment,
it is needed that the count query only counts events that have a
deployment. Otherwise, the pagination.total value in the response won't
be consistent with the number of elements in items.

* CFY-6332 Take into account the range filter for the timestamp

* CFY-6332 Remove unused test cases

The following is no longer required:
- _include set to None
- sort by timestamp

* CFY-6332 Fix default parameters for the build select query test cases

Include is no longer a parameter, but range_filters is

* CFY-6332 Fix event mapping test cases

Include parameter needs to be passed and `@timestamp` is now part of the
response.

* CFY-6332 Update docstrings

* CFY-6332 Use deployment filtering in the count query

When counting the number of events that match a query, take the
deployment filtering into account.

* CFY-6332 Filter by range in the count query

When a timestamp range is passed, the count query needs to take it into
account.

* CFY-6332 Add missing range_filters paremeter to build_count_query method

* CFY-6332 Refactor range filtering code

A new method now takes care of filtering by range based on the filters
and the model being looked at.

* CFY-6332 Raise exception if filter by range field is not in the model

* CFY-6332 Fix dictionary items iteration

For some reason the `.items` method call was missing, but the for loop
expects not only the key, but also the value stored for that key.

* CFY-6332 Add _apply_sort method

This method applies sorting criteria and validates that the column to
sort by exists in the query

* CFY-6332 Fix method name

* CFY-6332 Refactor filtering into a single method

Filtering by execution/deployment is now performed by a method that is
reused in a few places instead of duplicating the same code.