CFY 6332. Fix search event message integration test (#784)
* Use SQL like operator syntax

Since Elasticsearch is not used and full text search has not been
enabled in Posgres, the SQL LIKE operator is used and the `message`
parameter is expected to follow its syntax.

* Remove test case that no longer applies

This test case was supposed to check that a query containing special
characters for Elasticsearch were properly escaped. Now that SQL LIKE
operator is used, it's expected that the user formats the `message`
parameter properly.