CFY 6479. Fix test execution logging (#800)
* Add JSONString database type

This is useful to make sure that a JSON object is stored as a string
without having to worry in how JSON objects are handled in different
databases (PostgreSQL vs. SQLite)

* Set error_causes type to JSONString

This is to make sure that results from the database will be already
python objects without having to manually deserialize them.

* Add error_causes to API response

Needed by the cloudify-cli to print this information if requested by
`cfy events list -v`

* Remove expect_traceback_wip from test case

The test case is now fixed, so there's no longer need for this flag.