CFY-6718 Add foreign key naming convention
This makes easier to work with foreign keys in migration scripts by
being able to autogenerate their names.

Note that the convention is then one that already uses PosgreSQL by
default, so this is just something that is convenient, but doesn't break
anything on a 4.0 database that wasn't generated by the migration
scripts.