CFY-6718 Use batch operations
This is to make migrations work in SQLite. Since it doesn't support
ALTER TABLE commands (except those to add a column), what a batch
operation does is:
- create a new table with all the constrains
- move the data over to the new table
- drop the old table
- rename the new one.

Note: in this commit some null constraints have been removed also for
compatibility with SQLite. The reason is that when a new column is
added, its default value is automatically set to NULL which is
incompatible with setting the column as NOT NULL. To fix this, what
needs to be done is:
- add the column
- update table content to avoid null values
- drop the null default
- add the not null constraint