CFY-7257 Use association object instead of secondary table
A secondary table works great when there no additonal columns to take
care of. . However, when that's the case, there's no way to access those
additional columns at the ORM level, which is something that will be
needed to set the role when adding/removing a user to a tenant.