storage_manager: don't break when used outside request context
When the storage_manager is used outside a request context, it throws
an error because it tries to automatically filter on the current
user. (only happens in list, not in eg. get)

This change makes it so that the user won't be checked if not run
from a request.

This will allow using the storage manager for more things (not just
getting) in cloudify-premium, and in the future, also in the admin cli.