Celery: don't send users/groups related to a tenant
Celery tasks don't require the users/groups related to a tenant, so
we don't need to send them.

If we did want to send them, we'd have to serialize them. Currently
the users and groups end up as sqlalchemy association proxy lists
of models, which can't be json-encoded directly.

Future work might involve whitelisting the attributes instead of
blacklisting, ie. creating a new dict there with just the few
attributes that are required, instead of serializing the whole
tenant object and removing the few attributes we don't want.