CFY-7257 Fix `Tenant.all_users` property to return a list
After the latest changes in the ORM, `all_users` was returning an
association proxy list which has side effects when the an element is
added to the list to keep relationships up-to-date.

However, the code that uses this property is not taking this into
account, so the way to fix this is to return a regular list to avoid the
undesired side effects.