storage_manager: check .is_resource before checking for admin
Authorization queries for a tenant, so we need to be able to look up
tenants BEFORE current_tenant is set.

It used to be that _add_permissions_filter checked for is_resource
before ever checking for administrator, so we should go back to that
behaviour, because now checking for administrator needs a tenant.

So this change simply moves the .is_resource check to BEFORE
touching current_tenant at all.
This also makes _add_permissions_filter be more similar to
_add_tenant_filter.