CFY-7805 When installing plugins, validate manager platform
When installing a plugin with executor: central_deployment_agent,
check that the plugin is valid for the manager's distro & platform.

This validation is only done if the yaml file was provided,
and in the yaml file, the executor for the plugin is
central_deployment_agent. If the yaml file was not provided, this check
is not performed.

If the plugin's supported platform/distro differs from the manager's,
an error is raised.