rabbitmq: provide the internal cert as cacertfile
RabbitMQ uses the 'cacertfile' setting as the chain to send with
the server certificate (in addition to using it for client
cert verification).

This means the certfile setting must contain the server cert, and
cacertfile must contain the intermediaries, if any.

In certfile, the additional intermediary is ignored, so we can use
the cloudify_internal_cert bundle there.
In cacertfile, the server file is ignored, so there we can also use
the same bundle.