amqp-postgres service: set RestartSec=2
Without a restart delay, the service is restarted many times in a short time
if the database is down. Then, systemd's burst limiting kicks in, and it stops
restarting it.

Instead, let's wait after each restart, to give the db some time to actually start.
Then systemd will keep restarting amqp-postgres until it actually starts, and it
will never limit it, because it restarts slowly enough.