amqp_postgres: when de-batching after IntegrityError, clear cache
It is possible that the IntegrityError is caused by the cache
being stale: in snapshot restore, the execution db-ID changes
after the db restore, even though cloudify-ID stays the same.

If that db-ID was already cached, we need to drop that item
from the cache.

To help this, let's just drop the whole cache on IntegrityErrors:
this happens rarely enough, that a more granular way isn't
required to keep overall performance high