Make sorting tests more stable
Just sorting by id doesn't guarantee that the ordering will be the
same, and so assertEqual tests can fail.
This is because sometimes items can have the same id, eg. we have
two nodes with id=http_server, for different blueprints. They can sort
differently if we don't order by the blueprint id as well.