Use PostgreSQL in unittests (#1548)
* Use postgresql in unittests

* Fixup _drop_db

- It uses delete, not drop - update the docstring
- Use .addCleanup
- session.remove() to avoid deadlocks

* ...do the same for the events test

* Use an actual user in the events test

Not a fake one, which of course will throw a foreign key constraint
error when inserting the other rows, because the user doesn't exist...

* Execution.status_display - add a cast

We have a `CASE WHEN` there, where one branch returns a
execution_status enum type, and the other branch returns a text type;
cast the execution_status to text, so that the CASE returns a text
always

* Make sorting tests more stable

Just sorting by id doesn't guarantee that the ordering will be the
same, and so assertEqual tests can fail.
This is because sometimes items can have the same id, eg. we have
two nodes with id=http_server, for different blueprints. They can sort
differently if we don't order by the blueprint id as well.

* amqp-postgres tests - no need for the postgres setup

It's now all done in the base test