Snapshot-related fixes (#1831)
* Integration tests: add manager to keep_tables

Don't drop managers between the tests, because they're not restored
then. The managers table need to be kept for the whole lifespan
of the test container

* Snap-res integration tests: increase timeout

The restore tests that install plugins might take longer than
70 seconds, especially if it's a plugin that requires pip installs
from the internet. Which is unfortunate that it does.

* Snapshot restore: don't die on missing stage metadata

In old snapshots (eg. 4.0) the metadata might be missing.
In that case, migrate down to the empty string ie. the oldest
migration

* Snap-res: ignore stage errors for snapshots older than 4.3

Those snapshots aren't always clearly restorable, and it's not
actually supported anyway. Simply ignore the errors so that the
snapshot can be at least partially restored, so the user gets their
cloudify data even if they lose their stage data

If the stage restore fails, make sure to upgrade stage db back to
current, instead of leaving it downgraded

* Snapshots: check current networks

The existing impl only really checked the preexisting manager
against itself.

Instead, actually check it against the snapshot.
And also store the active networks in the snapshot.

This partially reverts #1737