Resource authorization: fallback to resource_name permission
If there's no specific resource_list etc permissions, fall back to
just resource name. This is used for resources that don't need
specific list/get/update/create permissions, but are just using
a singlt overall permissions. Currently this is used only for
the 'operations' permission, which is for all operations/tasks graphs
actions.