Pause amqppostgres while restoring a snapshot [ci skip]
While the snapshot is being restored, the database is downgraded
and upgraded back, and the snapshot execution itself might not
be in the database during that time.
If we try to insert logs during that time, they won't be stored,
because they can't have a FK to a non-existent execution.
Pause amqp-postgres so that all the logs are only stored after
the database has finished restoring

because CI doesn't test snap-res anyway
[ci skip]