amqp-postgres: insert timestamp at utc (#1882)
This reverts #1504

We only use the TIMESTAMP WITHOUT TIME ZONE type, and those are not
affected by the timezone setting. However, the now() function is.

This means that when we used now() without the `AT TIME ZONE`
specifier, we inserted timestamps at local timezone in amqp-postgres,
however we did insert timestamps at utc in the restservice side, due
to #1493

The comment was wrong when it said that the "DB is changing": the
DB timezone setting doesn't affect neither inserts or selects from
TIMESTAMP WITHOUT TIME ZONE columns, it only affects the output of
the now() function (and it also affects TIMESTAMP WITH TIME ZONE
columns, but we don't use those).