integration tests test_plugins_sort: don't install the plugins
For sorting, installation is irrelevant, so don't do it.
By marking the wagon as requiring a platform different than the
manager (here: windows), the manager won't attempt to install it.

This also means the plugin-name-change mechanic (where we change the
plugin name from "x" to "installing-x" while it's being installed)
won't come into play, so this will prevent the test from failing
on an unexpected plugin name.