CY-2609 Use unicode wherever possible (#2151)
* SQLModelBase: also handle unicode

* replace basestring with text_type

* replace unicode with text_type

* base64: bytes & unicode

we _do_ actually want unicode everywhere, but python's base64
insists on requiring and returning bytes. So let's oblige.

* use bytes where applicable

in the few instances where we do work with (possibly binary) files,
do use bytes, explicitly

* sha256: pass bytes in

hashing is a binary operation, and it needs bytes

tokes are always an uuid, so ascii is good enough and strict enough

* use unicode where possible

in those cases, we don't _need_ to work with bytes, so we can use
unicode.

* remove commented-out line [ci skip]

* Run unittests on py3

Drop v1/v2 tests, drop tox. Run all kinds of tests in the same ci job.

* flake

* test_premium: install cloudify-premium as well