CY-2794 Set hostname to cloudify in IP setter
This is to work around a problem with very long instance names on openstack, which 'hostname -s'
then truncates. If it happens to truncate ending with a hyphen then we generate an invalid DNS
SAN in a cert.
It is set to cloudify in case a generated VM has another hostname set, so that we don't put the
original VM's name in the certs when IP setter runs.