CY-1915 Integration tests: use docker, not docl (#2220)
* CY-1915 Replace docl with straight docker calls

The mounting-code stuff is not supported as of this commit.
Maybe later it will be again. Maybe not. We'll see.

* CY-1915 Add a pytest flag for providing the image name

Allows parametrizing the image naem

* rename docl.py to docker.py

* remove docl from reqs.txt

* remove docl usage in tests themselves

* env.stop_dispatch_processes is not a classmethod

* re-add build_resource_mapping

it is actually required for mounting the plugins.

* rework docker.run_manager

mount a config file. Yes, with a hardcoded admin password, whatever.
Also, use the builtin wait-for-starter instead of an ad-hoc port
sleeping

* replace old docl interface with the new docker interface

use container_id, container_ip, etc

* add a jenkinsfile

* use wait_for_deployment_environment

* use rest-client

* test_events: run_query correctly

* update readme

* add pytest to install_requiress