Integration tests changes: move env to fixtures (#2235)
* Fully move creating the container into conftest

The env stuff is not used now. Container is created
in a fixture and attached to the class. It's still keeping
a similar interface (.env.container_id).

* conftest.py: fixturize mounting sources

All the sources mounting (yes we still keep that, it's kinda
useful for deving) is now in this place.
The targets for mounting are, for now, just the most useful ones;
when needed, we'll add more.
This is mostly moving the logic from env.py

* delete unused env.py

* container: skip_validations=true

we don't really need validations here. The container is
a known environment, there's no need to validate.
Specifically, there's no need to validate empty disk space.
My disks are full all the time, man, i still want to be able
to run the tests.

* Move the integration-plugin-storage to /tmp

There's no reason for it to be under /opt, and it being under
/opt would require some chmods

* Wait for only 2 minutes for the manager to start

260 seconds is too long. 2 minutes is enough. 1 minute is
usually enough, except when the host machine is under
very heavy load (eg. starting 3 of them at once)

* Drop the status-reporter stuff from reset_storage.py

* Increase execution timeout in workflow tests

We seem to have made the manager slower???
That's unfortunate. But anyway, i want those tests to pass.
We'll examine performance... "later"

* document prepare_manager_storage

* remove unused _retry

Co-authored-by: Łukasz Maksymczuk <lukasz@lukasz-maksymczuk.intra.codilime.com>