rewrite several tests to use runtime_properties instead of a file
plugins used to write some data into a file on the manager
filesystem, and the tests used to examine that file.

You might've noticed I've also changed the plugins to instead of
storing in the file, store data in runtime-props.

So now, change the tests to look into runtime-props
instead of in the files