integration tests updates: plugins, fixtures (#2248)
* storage_manager: don't import sqlite

we've been using sqlite in unittests yeeeears ago, but now we don't.
So don't import it. This makes us not require sqlite to be available.

* python 3 compat for integration tests

the usual decodes

* move several plugins to integration_tests_plugins

it used to be that the plugin code lives in one directory, the
yaml file lives in another, and the code is mounted into the
container.

Instead:
 - move yaml files next to the plugin code
 - make the plugin code be a proper package with a setup.py
 - make sure the thing lives in integration_tests_plugins/ and not
   in several different places

* package & upload plugins in fixtures

instead of mounting the plugin code:
 - for each used plugin, package it into a wagon at session start
 - upload it, function-scoped

* use the uploaded plugins in tests

* rewrite several tests to use runtime_properties instead of a file

plugins used to write some data into a file on the manager
filesystem, and the tests used to examine that file.

You might've noticed I've also changed the plugins to instead of
storing in the file, store data in runtime-props.

So now, change the tests to look into runtime-props
instead of in the files

* clean up a few fixtures

so yeah, THAT'S where upload_license will live.
And we don't delete the license between tests. Great.

* agent installation tests: repackage the agent

similar as it was before, repackage the agent with the given sources

* add mgmtworker and agent to mounts

* drop a test that was overdue for dropping

i'm not gonna go updating a 3.2 test. Let's just forget that it existed

* cleanup and simplify the dockercompute plugin

funnily, the other branch (install_agent=false) was never used

anyway, this is the Simple™ way of installing an agent: generate
the script, run it.

All plugin-based installs do a variation of this (eg. the openstack
plugin, who instead of executing it directly, puts it in userdata)