Drop invalid tests
Those tests kill-cancel a workflow, then resume it, and wait for
it to finish.
However, that workflow will never finish, because a resumed task
is not resent (after the no-double-execution change). For it to have
been resent, the test should have done "force-resume", aka
`cfy exec resume --reset-operations`. And we already have tests
doing that.