Resume integration test updates (#2262)
* Drop invalid tests

Those tests kill-cancel a workflow, then resume it, and wait for
it to finish.
However, that workflow will never finish, because a resumed task
is not resent (after the no-double-execution change). For it to have
been resent, the test should have done "force-resume", aka
`cfy exec resume --reset-operations`. And we already have tests
doing that.

* Task state is response-sent after a response has been sent.

It's only succeeded after the response has been received. In that
test, the execution is cancelled, and is not resumed, so the
operation response is still waiting in the queue. Not succeeded yet.
It will be succeeded as soon as the workflow is resumed.

* pass ip to create_rest_client

* failing task test: throw nonrecoverable

we don't want retries here. This is for test_resume_failed