Kill-cancel: bypass maintenance
When kill-cancelling, always bypass maintenance. Killing stuff
is as forceful as we can get, so we need to bypass everything.

And it's useful, because you mostly only kill things when trying
to recover a manager that is in an invalid state; and recovering
managers is what maintenance mode is for