Delete PluginsTest and friends
This was supposed to be some kinda base class to use in plugin
integration tests, but it's not used, because the plugins
now use the ecosystem_tests from cloudify-incubator.

Drop the whole thing and its friends.

(if this turns out to be needed, feel free to revert this commit,
but if you do, fix this thing to do ABSOLUTELY NO HTTP REQUESTS AT
IMPORT TIME)