CY-3302 Plugin uninstall fixes (#2302)
* Plugin.installation_state: declare cascade=delete

Tell the ORM that this FK is indeed cascading on deletes, and that
it doesn't need to do anything when deleting a plugin, because
the DB will take care of it.

* Plugin uninstall: skip useless check

plugin_installable_on_current_platform is not used when installing,
so it shouldn't used when uninstalling either

* delete plugin_installable_on_current_platform

not used anymore

* 2.1 plugins: drop useless try/excepts

- `except ManagerException: raise` is obviously useless
- `except Exception: reraise` is bullshit, just show the original
  exception instead
- `except ExecutionTimeout` is just a misunderstanding of how this
  works, the restservice doesn't wait for executions to finish,
  so it will never get a timeout. It only runs the execution.
  Well, used to, now it's a service task and not an execution anyway

* drop test_uninstall_timeout

you gotta love that patches a function to throw, and then asserts
that it did throw. Anyway, 2 commits ago I removed that except,
and so removing the test now

* drop test_uninstall_failure

no, i removed the reraise-as-PluginInstallationError.
Unhandled/unexpected exceptions are a 500. That's it.

Plugin remove is supposed to not throw anything, if it does, it's
a 500. Amen.