Snapshot restore: don't worry about plugins (#2314)
* snapshot_restore: remove the plugin bits

we're now not installing plugins up-front, so remove all the
things related to plugins

* restservice: stop passing ignore_plugin_failure to snap-res

that arg is now unused, so no need to pass it in.

* create_dep_env: drop unused bits

* test_snapshot: only have one "bad plugin" test

we don't install plugins in snap-res. No need to test multiple
kinds of failure there. It doesn't fail at all. It might fail
when you force-install or try to use the plugin, of course,
but not in snap-res.