CY-3320 Delete deployment after deleting dep env
Deleting the deployment from inside the workflow means that
the execution is deleted (via cascade from deployment) while it is
still running.
Deleting executions that are running leads them to fail.

This means deleting deployments from inside delete-dep-env
is racy, because if it just happens to poll the execution state
(to see if it was cancelled) after it was already deleted, it's
going to get a 404 and break.

Instead, delete the deployment after the delete-dep-env workflow
has finished.