CY-3320 Delete deployment after deleting dep env (#2320)
* CY-3320 Delete deployment after deleting dep env

Deleting the deployment from inside the workflow means that
the execution is deleted (via cascade from deployment) while it is
still running.
Deleting executions that are running leads them to fail.

This means deleting deployments from inside delete-dep-env
is racy, because if it just happens to poll the execution state
(to see if it was cancelled) after it was already deleted, it's
going to get a 404 and break.

Instead, delete the deployment after the delete-dep-env workflow
has finished.

* only on terminated

* deployment-environment: simplify the workflows

no need to create a tasks graph to send an event. Really. Just make
it a log.

* dep-env: remove unused stuff

* delete deployment: remove the directory in resource-manager

no need to do it in the endpoint, if we're already using
resource-manager as the controller layer, speaking mvc-ish

* refactors

1. delete_deployment -> delete_deployment_environment
2. factor out the "really delete deployment" parts from update-exec
   into delete_deployment

* Fixup unittests