test_amqp_queues_list: re-fetch queues/exchanges for assertions
we are going to be asserting on queues/exchanges after uninstalling
the agent, so we must fetch the list of queues/exchanges after
uninstalling the agent.

We used to uninstall the agent, and then assert on the list of
exchanges that we fetched while the agent was still installed.

I think it was a pre-merge cleanup gone wrong.