CY-3420 Passive_deletes for executions (#2331)
* Pass `backref` directly into relationships

Instead of passing just the name, allow passing the whole object.
This will allow greater configurability, because now I can actually
configure the backref when creating a one_to_many_relationship

* Set passive_deletes for execution.deployment's backref

When a deployment is deleted, the db cascades the execution
by itself. No need to do it on the ORM side.