CY-3420 Lightweight deployments delete
Deleting deployments can lead to fetching an unbounded amount of
items - executions and node instances - to delete it ORM-side, and to
return them as part of the response.

Instead of deleting them ORM-side, delete them DB-side: we already
have the CASCADE set up on the DB side, we just need to say
passive_deletes so that sqlalchemy knows not to touch the
relationship.

Instead of returning the whole deployment object, with embedded
relationships, just return a mock object with just the id.
We can't avoid returning _something_ because the 4.6 CLI expects us
to return a dict.
We'll only do it for deployments, because this is the most affected
endpoint, and other endpoints aren't having problems in practice
(and this is a 4.6 fix, on an as-needed basis)

Also, don't fetch all of the node-instances. Only look for the
ones that are actually not uninitialized/deleted. And only the
first page of those, we don't need to spew out 100k node instance ids
if we are going to fail anyway.