deployment delete: return a rendered execution object
when running PATCH to update the delete-dep-env execution status,
which makes the deployment be actually deleted, that makes the
execution be deleted as well, and the request returns 500.

Instead, pre-render the execution to-be-returned ahead of time,
before we delete it.

It is a bit ugly to return a rendered dict from here, but we don't
really have a better way.

A long-term better way would be to not ad-hoc'ly get/render
attributes from the database in rendering, but only fetch them
once. But that's a framework-level question.