Error logging: make auth errors DEBUG
All kinds of bad auth can flood the logs with useless logs.
The big offender is the status checkers, as even the built-in one
generates one log every 15 seconds, and of course there can be
user-specific ones on top of that.

These logs are also useless for detecting attacks, as they don't
include any kind of identification with them (eg. an IP). Nginx logs
serve better for that.

Hide the auth logs by default, by making them debug-level.