Config: list rabbitmqbroker columns explicitly
Instead of loading the whole model, load only the columns
we need.
This means that we won't try to load columns that aren't needed,
eg. .monitoring_username (and others).
This will allow a 5.1 restservice to at least load the rabbitmq
brokers when faced with a 4.x db (which happens in snapshot-restore)