Plugin update: cope with plugin being deleted concurrently
If this update runs at the same time as deleting the plugin
(which will happen, since we delete a plugin, and at the same time
agents delete its directory and attempt to update the state to
"uninstalled"), then we will see a race between updating the state
and deleting the rows.

If update happens first, great, the row is deleted later.

If delete happens first, the .get() will throw a 404 and all's good.

If however the delete happens between the .get and the update,
then we get an error; in this case, let's confirm that the
plugin is indeed deleted, and throw a 404 as expected.

Getting a 'stale' error in other cases than the plugin being
already deleted, should never happen, and will still be a 500.