Allow for more than one process to be reported
With the recent transition to recording rules, status reporter can have
the same system service process reported more than once (because of the
`sum by` aggregation used in recording rules).  While this might be a
little bit controversial and could mean some mis-configuration, it is
perfectly safe to assume values `>= 1` mean process is running (just
as `== 1` indicated it when simple queries were used).