RD-664 Allow using alembic directly to migrate (#2258)
* Alembic boilerplate to be able to use migrations directly

Alembic requires an alembic.ini to use the alembic cli.

Also, update the env.py to load the necessary stuff so that we
can use alembic directly and not flask-migrate

* sql-friendly way of creating enums

just .create(bind) uses the ORM, and is not compatible with offline
migrations

* Replace straight values with inline_literal

This way sqlalchemy knows it doesn't need to adapt the values,
and doesn't need actual db connection to do it. It's just a literal.

* Replace orm-side inserts/deletes with sql-side ones

This approach is compatible with offline.