inte-tests: log `docker run` command
It's useful to log the `docker run` call because then developers can
copy it and use it for running their containers manually,
with all the mounts formatted in. I do it all the time.

With pytest, this log is not going to be shown by default,
but you can `--log-cli-level=info` (or pytest.ini).