RD-653 Identify blueprints for a removed plugin
A fix in an algorithm that chooses which blueprints would be affected
by a plugin removal.  Deployments are not taken into considerations,
blueprint-plugins relationships are based on blueprint plans solely.

Also in case a plugin is used in any blueprint and hence should not be
removed, the list printed out as an error message contains valid
blueprint ids.